Feat: Implement to FindCommand to search tasks with keyword
We lack a way to search and filter tasks via keywords, a function that is essential for users to find specific tasks.

Let's implement a FindCommand class that expects user input in the form of 'find <keyword>', searches through the existing task list and returns a new list of matching tasks. If there are no matching tasks found, the user is informed via a UI message too.

Other fixes: Added UnknownCommand class for error handling when parser does not recognise command given